Reorganize tests by splitting them out into appropriate modules.
This will make it easier to rip out hail1 when the time comes.
